new_requirements.txt
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
et_xmlfile==2.0.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pynput==1.7.7
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
python-dotenv==1.0.1
python-xlib==0.33
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
termcolor==2.5.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
wsproto==1.2.0
